feat: show task cards
